+++ pkg/plugin/query.go
@@ -53,7 +53,7 @@ type Query struct {
 	IntervalMs   int64     `json:"intervalMs"`
 	MaxLines     int       `json:"maxLines"`
 	Step         string    `json:"step"`
-	Field        string    `json:"field"`
+	Fields       []string  `json:"field"`
 	QueryType    QueryType `json:"queryType"`
 	url          *url.URL
 	ForAlerting  bool `json:"-"`
@@ -256,7 +256,9 @@ func (q *Query) histQueryURL(queryParams url.Values, minInterval time.Duration) 
 	values.Set("start", strconv.FormatInt(q.TimeRange.From.Unix(), 10))
 	values.Set("end", strconv.FormatInt(q.TimeRange.To.Unix(), 10))
 	values.Set("step", step)
-	values.Set("field", q.Field)
+	for _, f := range q.Fields {
+		values.Add("field", f)
+	}
 
 	q.url.RawQuery = values.Encode()
 	return q.url.String()